kern/efi/mm: try allocating in the lower 4GB first
On arm64 we can't restrict memory usage below 0xffffffff as we do
on x86 because there is no guarantee memory in that range exists
at all (an example is Apple Silicon).
Instead, try to allocate in the lower 4GB range first and fall
back to the entire available address space if that did not work.

This fixes a bug on Snapdragon X Elite devices such as the Lenovo
Thinkpad T14s which crashes when accessing memory above 0xffffffff.

Signed-off-by: Tobias Heider <tobias.heider@canonical.com>