appended signatures: Support verifying appended signatures
Building on the parsers and the ability to embed X.509 certificates, as
well as the existing gcrypt functionality, add a module for verifying
appended signatures.

This includes a verifier that requires that Linux kernels and
GRUB modules have appended signatures, and commands to manage the
list of trusted certificates for verification.

Verification must be enabled by setting check_appended_signatures. If
secure boot is enabled with enforced mode when the module is loaded,
verification will be enabled and locked automatically. If verification
is enabled, extract trusted keys from the GRUB ELF Note and store them in the db.

As with the PGP verifier, it is not a complete secure-boot solution:
other mechanisms, such as a password or lockdown, must be used to ensure
that a user cannot drop to the GRUB shell and disable verification.

Introducing the following GRUB commands.

 1. append_list_db:
      Show the list of trusted certificates from the db list
 2. append_add_db_cert:
      Add the trusted certificate to the db list
 3. append_rm_dbx_cert:
      Remove the distrusted certificate from the db list
 4. append_verify:
      Verify the signed file using db list

Signed-off-by: Daniel Axtens <dja@axtens.net>
Signed-off-by: Sudhakar Kuppusamy <sudhakar@linux.ibm.com>
Reviewed-by: Stefan Berger <stefanb@linux.ibm.com>
Reviewed-by: Avnish Chouhan <avnish@linux.ibm.com>