appended signatures: Read default db keys from the ELF Note
If Secure Boot is enabled with dynamic key management mode and the
use_static_keys flag is not set, then read the keys from the PKS
keystore as well from ELF Note, and add them into the db.
Otherwise, it only reads the static keys from the ELF Note and
adds them into the db.

Signed-off-by: Sudhakar Kuppusamy <sudhakar@linux.ibm.com>
Reviewed-by: Stefan Berger <stefanb@linux.ibm.com>
Reviewed-by: Avnish Chouhan <avnish@linux.ibm.com>